server/requirements.txt
beautifulsoup4
fastapi
fitz
huggingface_hub

langchain_core
langchain_groq
langchain_pinecone
langchain_text_splitters
langchain_cloudflare

pinecone

pdf2docx
pydantic
pymupdf4llm
python-dotenv
